Refactor CRE system-tests
Refactoring: move re-used common utility methods in the tests to smoke-tests/cmd package

Refactoring: move re-used logically grouped steps to helpers_test.go

Refactoring: externalize CREv1 PoR smoke-test to a separate file.

Refactoring: externalize CREv1 PoR smoke-test to a separate file.

Refactoring: externalize CREv2 smoke-tests to separate files.

Refactoring: update system-test files names and README.
