+++ .changeset/red-cats-attend.md
@@ -1,0 +1,5 @@
+---
+"chainlink": minor
+---
+
+#internal Refactor system tests
