+++ system-tests/lib/cre/contracts/contracts.go
@@ -409,6 +409,29 @@ func MustFindAddressesForChain(addressBook cldf.AddressBook, chainSelector uint6
 	return addr
 }
 
+// MergeAllDataStores merges all DataStores (after contracts deployments)
+func MergeAllDataStores(fullCldEnvOutput *cre.FullCLDEnvironmentOutput, changesetOutputs ...cldf.ChangesetOutput) {
+	fmt.Print("Merging DataStores (after contracts deployments)...")
+	minChangesetsCap := 2
+	if len(changesetOutputs) < minChangesetsCap {
+		panic(fmt.Errorf("DataStores merging failed: at least %d changesets required", minChangesetsCap))
+	}
+
+	// Start with the first changeset's data store
+	baseDataStore := changesetOutputs[0].DataStore
+
+	// Merge all subsequent changesets into the base data store
+	for i := 1; i < len(changesetOutputs); i++ {
+		otherDataStore := changesetOutputs[i].DataStore
+		mergeErr := baseDataStore.Merge(otherDataStore.Seal())
+		if mergeErr != nil {
+			panic(errors.Wrap(mergeErr, "DataStores merging failed"))
+		}
+	}
+
+	fullCldEnvOutput.Environment.DataStore = baseDataStore.Seal()
+}
+
 func ConfigureWorkflowRegistry(testLogger zerolog.Logger, input *cre.WorkflowRegistryInput) (*cre.WorkflowRegistryOutput, error) {
 	if input == nil {
 		return nil, errors.New("input is nil")
@@ -518,4 +541,62 @@ func ConfigureDataFeedsCache(testLogger zerolog.Logger, input *cre.ConfigureData
 	input.Out = out
 
 	return out, nil
+}
+
+func DeployDataFeedsCacheContract(testLogger zerolog.Logger, chainSelector uint64, fullCldEnvOutput *cre.FullCLDEnvironmentOutput) (common.Address, cldf.ChangesetOutput, error) {
+	testLogger.Info().Msg("Deploying Data Feeds Cache contract...")
+	deployDfConfig := df_changeset_types.DeployConfig{
+		ChainsToDeploy: []uint64{chainSelector},
+		Labels:         []string{"data-feeds"}, // label required by the changeset
+	}
+
+	dfOutput, dfErr := commonchangeset.RunChangeset(df_changeset.DeployCacheChangeset, *fullCldEnvOutput.Environment, deployDfConfig)
+	if dfErr != nil {
+		return common.Address{}, cldf.ChangesetOutput{}, errors.Wrapf(dfErr, "failed to deploy Data Feeds Cache contract on chain %d", chainSelector)
+	}
+
+	mergeErr := fullCldEnvOutput.Environment.ExistingAddresses.Merge(dfOutput.AddressBook) //nolint:staticcheck // won't migrate now
+	if mergeErr != nil {
+		return common.Address{}, cldf.ChangesetOutput{}, errors.Wrap(mergeErr, "failed to merge address book of Data Feeds Cache contract")
+	}
+	testLogger.Info().Msgf("Data Feeds Cache contract deployed to %d", chainSelector)
+
+	dataFeedsCacheAddress, dataFeedsCacheErr := FindAddressesForChain(
+		fullCldEnvOutput.Environment.ExistingAddresses, //nolint:staticcheck // won't migrate now
+		chainSelector,
+		df_changeset.DataFeedsCache.String(),
+	)
+	if dataFeedsCacheErr != nil {
+		return common.Address{}, cldf.ChangesetOutput{}, errors.Wrapf(dataFeedsCacheErr, "failed to find Data Feeds Cache contract address on chain %d", chainSelector)
+	}
+	testLogger.Info().Msgf("Data Feeds Cache contract found on chain %d at address %s", chainSelector, dataFeedsCacheAddress)
+
+	return dataFeedsCacheAddress, dfOutput, nil
+}
+
+func DeployReadBalancesContract(testLogger zerolog.Logger, chainSelector uint64, fullCldEnvOutput *cre.FullCLDEnvironmentOutput) (common.Address, cldf.ChangesetOutput, error) {
+	testLogger.Info().Msg("Deploying Read Balances contract...")
+	deployReadBalanceRequest := &keystone_changeset.DeployRequestV2{ChainSel: chainSelector}
+	rbOutput, rbErr := keystone_changeset.DeployBalanceReaderV2(*fullCldEnvOutput.Environment, deployReadBalanceRequest)
+	if rbErr != nil {
+		return common.Address{}, cldf.ChangesetOutput{}, errors.Wrap(rbErr, "failed to deploy Read Balances contract")
+	}
+
+	mergeErr2 := fullCldEnvOutput.Environment.ExistingAddresses.Merge(rbOutput.AddressBook) //nolint:staticcheck // won't migrate now
+	if mergeErr2 != nil {
+		return common.Address{}, cldf.ChangesetOutput{}, errors.Wrap(mergeErr2, "failed to merge address book of Read Balances contract")
+	}
+	testLogger.Info().Msgf("Read Balances contract deployed to %d", chainSelector)
+
+	readBalancesAddress, readContractErr := FindAddressesForChain(
+		fullCldEnvOutput.Environment.ExistingAddresses, //nolint:staticcheck // won't migrate now
+		chainSelector,
+		keystone_changeset.BalanceReader.String(),
+	)
+	if readContractErr != nil {
+		return common.Address{}, cldf.ChangesetOutput{}, errors.Wrap(readContractErr, "failed to find Read Balances contract address")
+	}
+	testLogger.Info().Msgf("Read Balances contract found on chain %d at address %s", chainSelector, readBalancesAddress)
+
+	return readBalancesAddress, rbOutput, nil
 }
