+++ system-tests/lib/cre/workflow/docker.go
@@ -43,7 +43,23 @@ func findAllDockerContainerNames(pattern string) ([]string, error) {
 	return containerNames, nil
 }
 
-func CopyArtifactToDockerContainers(filePath string, containerNamePattern string, targetDir string) error {
+func CopyArtifactsToDockerContainers(containerTargetDir string, containerNamePattern string, filesToCopy ...string) error {
+	for _, file := range filesToCopy {
+		if _, err := os.Stat(file); err != nil {
+			fmt.Fprintf(os.Stderr, "Warning: File '%s' does not exist. Skipping file copying to docker containers\n", file)
+			continue
+		}
+
+		workflowCopyErr := copyArtifactToDockerContainers(file, containerNamePattern, containerTargetDir)
+		if workflowCopyErr != nil {
+			return errors.Wrapf(workflowCopyErr, "failed to copy a file (%s) to docker containers", file)
+		}
+	}
+	return nil
+}
+
+func copyArtifactToDockerContainers(filePath string, containerNamePattern string, targetDir string) error {
+	fmt.Printf("Copying file '%s' to Docker containers.\n", filePath)
 	containerNames, containerNamesErr := findAllDockerContainerNames(containerNamePattern)
 	if containerNamesErr != nil {
 		return errors.Wrap(containerNamesErr, "failed to find Docker containers")
