+++ system-tests/tests/smoke/cre/cre_suite_test.go
@@ -1,0 +1,59 @@
+package cre
+
+import (
+	"testing"
+	"time"
+)
+
+// TODO: Where to move these constants? Is it appropriate place for keeping defaults here?
+const (
+	AuthorizationKeySecretName = "AUTH_KEY"
+	// TODO: use once we can run these tests in CI (https://smartcontract-it.atlassian.net/browse/DX-589)
+	// AuthorizationKey           = "12a-281j&@91.sj1:_}"
+	AuthorizationKey = ""
+
+	// Test configuration constants
+	DefaultVerificationTimeout = 5 * time.Minute
+	DefaultRetryInterval       = 2 * time.Second
+	DefaultValidationInterval  = 10 * time.Second
+)
+
+/*
+To execute tests locally start the local CRE first:
+Inside `core/scripts/cre/environment` directory
+ 1. Ensure the necessary capabilities (i.e. readcontract, http-trigger, http-action) are listed in the environment configuration
+ 2. Run: `go run . env start && ctf obs up && ctf bs up` to start env + observability + blockscout.
+ 3. Execute the tests in `system-tests/tests/smoke/cre`: `go test -timeout 15m -run ^Test_CRE_Suite$`.
+*/
+func Test_CRE_Suite(t *testing.T) {
+	testEnv := SetupTestEnvironment(t)
+
+	// WARNING: currently we can't run these tests in parallel, because each test rebuilds environment structs and that includes
+	// logging into CL node with GraphQL API, which allows only 1 session per user at a time.
+	t.Run("[v1] CRE Suite", func(t *testing.T) {
+		// requires `readcontract`, `cron`
+		t.Run("[v1] CRE Proof of Reserve (PoR) Test", func(t *testing.T) {
+			ExecutePoRTest(t, testEnv)
+		})
+	})
+
+	t.Run("[v2] CRE Suite", func(t *testing.T) {
+		t.Run("[v2] vault DON test", func(t *testing.T) {
+			ExecuteVaultTest(t, testEnv)
+		})
+
+		t.Run("[v2] HTTP trigger and action test", func(t *testing.T) {
+			// requires `http_trigger`, `http_action`
+			ExecuteHTTPTriggerActionTest(t, testEnv)
+		})
+
+		t.Run("[v2] DON Time test", func(t *testing.T) {
+			const skipReason = "Implement smoke test - https://smartcontract-it.atlassian.net/browse/CAPPL-1028"
+			t.Skipf("Skipping test for the following reason: %s", skipReason)
+		})
+
+		t.Run("[v2] Beholder test", func(t *testing.T) {
+			ExecuteBeholderTest(t, testEnv)
+		})
+	})
+}
